Moved ReferenceDumper to Dumper\YamlReferenceDumper
This is done to make other formats, such as the XmlReferenceDumper
possible. This deprecates the old
Symfony\Component\Config\Definition\ReferenceDumper.
